refactor : 이벤트 신청서 entity 불필요 필드 삭제 후 에러 생기는 부분 각주 처리 + user entity 패키지 구조 정리
