+++ run/src/main/java/com/guide/run/event/service/AllEventGetService.java
@@ -26,7 +26,7 @@ import java.util.List;
 public class AllEventGetService {
     private final EventRepository eventRepository;
     private final EventFormRepository eventFormRepository;
-
+/*
     public List<AllEvent> getAllEvent(int start, int limit, String sort, String type, String kind, String privateId) {
         Pageable pageable = PageRequest.of(start/limit,limit);
         List<AllEvent> events = new ArrayList<>();
@@ -347,5 +347,5 @@ public class AllEventGetService {
             default:
                 throw new NotValidKindException();
         }
-    }
+    }*/
 }
