+++ run/src/main/java/com/guide/run/global/security/user/CustomUserDetailsService.java
@@ -2,7 +2,7 @@ package com.guide.run.global.security.user;
 
 import com.guide.run.global.exception.user.resource.NotExistUserException;
 import com.guide.run.user.entity.user.User;
-import com.guide.run.user.repository.UserRepository;
+import com.guide.run.user.repository.user.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
