+++ run/src/main/java/com/guide/run/user/controller/TestController.java
@@ -4,7 +4,7 @@ package com.guide.run.user.controller;
 import com.guide.run.global.exception.user.resource.NotExistUserException;
 import com.guide.run.user.entity.type.Role;
 import com.guide.run.user.entity.user.User;
-import com.guide.run.user.repository.UserRepository;
+import com.guide.run.user.repository.user.UserRepository;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.CrossOrigin;
